% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-density.R
\name{dtailweibull}
\alias{dtailweibull}
\title{Density function}
\usage{
dtailweibull(x, threshold, scale, shape)
}
\arguments{
\item{x}{Sample data.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\description{
This function computes the density of the tail weibull with the input sample data. The expression for the density used is:
\deqn{abx^{(b-1)}\exp(-ax^{b}+au^{b})}
}
\examples{
# Generate random deviates from a weibull tail and plot the theoretical density.
scale <- 2
shape <- 1
threshold <- 1
x_seq <- seq(threshold, 5, length.out = 500)
theo_density <- dtailweibull(x_seq, threshold = threshold, scale = scale, shape = shape)
sample <- rtailweibull(500, threshold = threshold, scale = scale, shape = shape)
hist(sample, probability = T)
lines(x = x_seq, y = theo_density, col = "red")
}
\keyword{weibull}
