% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-mle.R
\name{fittailw}
\alias{fittailw}
\title{Maximum Likelihood Estimation of Weibull tails}
\usage{
fittailw(sample)
}
\arguments{
\item{sample}{Sample data.}
}
\description{
This function scales the input data w.r.t. the threshold and performs MLE with a Weibull tail.
}
\examples{
scale <- 2
shape <- 1
threshold <- 1
s <- rtailw(1000, threshold = threshold , scale = scale, shape = shape)
fits <- fittailw(s)
x_seq <- seq(threshold, max(s), length.out = 500)
theo_density <- dtailw(x_seq, threshold = threshold, scale = fits$scale, shape = fits$shape)
hist(s, probability = TRUE, breaks = "FD")
lines(x = x_seq, y = theo_density, col = "red")
}
\keyword{weibull}
