% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-loglikelihood.R
\name{ltailw}
\alias{ltailw}
\title{TailW Log-likelihood function}
\usage{
ltailw(x, threshold, scale, shape)
}
\arguments{
\item{x}{Sample data.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\value{
Gives the log-likelihood of the TailW. The length of the result is determined by the length of x.
}
\description{
This function computes the log-likelihood of the tailW with the input sample data.
\deqn{l(x;\alpha,\beta) = n(\log(\alpha)+\log(\beta))+(\beta-1)\sum_{i=1}^{n}\log(x+\nu)-\alpha\sum_{i=1}^{n}((x+\nu)^\beta-\nu^\beta)}
}
\examples{
ltailw(1,1,1,1)
}
\keyword{TailW}
