% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-loglikelihood.R
\name{ltailweibull}
\alias{ltailweibull}
\title{Log-likelihood function}
\usage{
ltailweibull(x, threshold, scale, shape)
}
\arguments{
\item{x}{Sample data.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\description{
This function computes the log-likelihood of the tail weibull with the input sample data.
}
\examples{
ltailweibull()
}
\keyword{weibull}
