% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-probability.R
\name{ptailweibull}
\alias{ptailweibull}
\title{Probability function}
\usage{
ptailweibull(x, threshold, scale, shape)
}
\arguments{
\item{x}{Sample data.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\description{
This function computes the probability of the tail weibull with the input sample data.
}
\examples{
# Using the probability function to show the fitting.
samp <- rtailweibull(1000, 1, 2, 3)
emp_cdf <- ecdf(samp)(samp)
pars <- fittailweibull(samp)
x_seq <- seq(min(samp), max(samp), length.out = 250)
p <- ptailweibull(x_seq, threshold = 1, scale = pars$scale, shape = pars$shape)
plot(samp, 1-emp_cdf, log = "y")
lines(x_seq, 1-p, col = "red")
}
\keyword{weibull}
