% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG-random.R
\name{rFTG}
\alias{rFTG}
\title{FTG Random Sample Generation}
\usage{
rFTG(n, threshold, scale, shape)
}
\arguments{
\item{n}{Sample size.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\value{
Gives random deviates of the FTG. The length of the result is determined by n.
}
\description{
This function computes n random variates from full-tail gamma with a rejection method.
}
\examples{
x <- rFTG(100, 1, 1, 1)
hist(x, breaks = "FD")
}
\references{
del Castillo, Joan & Daoudi, Jalila & Serra, Isabel. (2012). The full-tails gamma distribution applied to model extreme values. ASTIN Bulletin. <doi:10.1017/asb.2017.9>.
}
\keyword{FTG}
