% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-random.R
\name{rtailweibull}
\alias{rtailweibull}
\title{Random function}
\usage{
rtailweibull(n, threshold, scale, shape)
}
\arguments{
\item{n}{Sample size.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\description{
This function generates random deviates for the tail weibull distribution.
}
\examples{
#rtailweibull()
}
\keyword{weibull}
