% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG-density.R
\name{dFTG}
\alias{dFTG}
\title{Density function}
\usage{
dFTG(x, alpha, theta, rho, threshold)
}
\arguments{
\item{x}{Sample data.}

\item{alpha}{alpha.}

\item{theta}{theta.}

\item{rho}{rho.}

\item{threshold}{Minimum value of the tail.}
}
\description{
This function computes the density of the full-tail gamma with the input sample data. The expression for the density used is:
\deqn{f(x; \alpha, \theta, \rho, u) = \theta(\rho + \theta(x + u))^{\alpha - 1}\exp(-(\rho + \theta(u + x)))/ \Gamma(\alpha, \rho + \theta u)}
}
\examples{
a <- 0.3
t <- 0.3
r <- 0.8
n <- 1000
sample <- rFTG(n, a, t, r)
sample <- sort(sample)
u <- sample[length(sample) - n/2]
xu <- seq(u, max(sample), length.out = 200)
du <- dFTG(xu, a, t, r, u)
hist(sample, breaks = "FD", probability = TRUE)
lines(xu, du, col = "red")
}
\keyword{dFTG}
