% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-mle.R
\name{fittailweibull}
\alias{fittailweibull}
\title{Maximum Likelihood Estimation of weibull tails}
\usage{
fittailweibull(sample)
}
\arguments{
\item{sample}{Sample data.}
}
\description{
This function scales the input data w.r.t. the threshold and performs MLE with a weibull tail.
}
\examples{
fittailweibull()
}
\keyword{weibull}
