% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG-loglikelihood.R
\name{lFTG}
\alias{lFTG}
\title{Loglikelihood function}
\usage{
lFTG(x, alpha, theta, rho)
}
\arguments{
\item{x}{Sample data.}

\item{alpha}{alpha.}

\item{theta}{theta.}

\item{rho}{rho.}
}
\description{
Loglikelihood function
}
\examples{
#lFTG()
}
\keyword{lFTG}
