% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG-loglikelihood.R
\name{lFTG}
\alias{lFTG}
\title{Loglikelihood function}
\usage{
lFTG(x, alpha, theta, rho)
}
\arguments{
\item{x}{Sample data.}

\item{alpha}{alpha.}

\item{theta}{theta.}

\item{rho}{rho.}
}
\description{
This function computes the loglikelihood of the full-tail gamma with the input sample data. The expression used is:
\deqn{l(x; \alpha, \sigma, \rho) = -n(\log\Gamma(\alpha, \rho) + \log(\sigma\rho^{-\alpha}) - \frac{\alpha - 1}{n}\sum_{i = 1}{n}\log(1 + \frac{x_{i}}{\sigma}) + \frac{\rho}{n} \sum_{i = 1}{n}(1 + \frac{x_{i}}{\sigma})}
}
\examples{
#lFTG()
}
\keyword{lFTG}
