% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG-probability.R
\name{pFTG}
\alias{pFTG}
\title{Probability function}
\usage{
pFTG(x, alpha, theta, rho)
}
\arguments{
\item{x}{Sample data.}

\item{alpha}{alpha.}

\item{theta}{theta.}

\item{rho}{rho.}
}
\description{
This function computes the probability of the full-tail gamma with the input sample data. The expression for the probability used is:
\deqn{G(x; \alpha, \theta, \rho) = 1 - \Gamma(\alpha, \rho + \theta x)/\Gamma(\alpha, \rho).}
}
\examples{
#pFTG()
}
\keyword{dFTG}
