% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-probability.R
\name{ptailweibull}
\alias{ptailweibull}
\title{Probability function}
\usage{
ptailweibull(x, threshold, scale, shape)
}
\arguments{
\item{x}{Sample data.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\description{
This function computes the probability of the tail weibull with the input sample data.
}
\examples{
ptailweibull()
}
\keyword{weibull}
