% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG_quantile.R
\name{qFTG}
\alias{qFTG}
\title{Probability function}
\usage{
qFTG(p, threshold, scale, shape)
}
\arguments{
\item{p}{probabilities.}

\item{threshold}{Threshold.}

\item{scale}{Scale.}

\item{shape}{Shape.}
}
\description{
This function computes the quantiles of the full-tail gamma with the input sample data. The expression for the probability used is:
\deqn{Q(p; \alpha, \theta, \rho) =\Gamma^{-1} \Big( \alpha, (1-p)\Gamma(\alpha, \rho) - \rho \Big)/ \theta }
}
\examples{
qFTG(0.5,1,1,1)
}
\keyword{dFTG}
