% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG_quantile.R
\name{qFTG}
\alias{qFTG}
\title{Probability function}
\usage{
qFTG(p, alpha, theta, rho)
}
\arguments{
\item{p}{probabilities.}

\item{alpha}{alpha.}

\item{theta}{theta.}

\item{rho}{rho.}
}
\description{
This function computes the quantiles of the full-tail gamma with the input sample data. The expression for the probability used is:
\deqn{F(x; \alpha, \theta, \rho) = 1- \frac{\Gamma(\alpha, \rho + \theta x)}{\Gamma(\alpha, \rho)}}
}
\examples{
#qFTG()
}
\keyword{dFTG}
