% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-quantile.R
\name{qtailw}
\alias{qtailw}
\title{Quantile function}
\usage{
qtailw(p, threshold, scale, shape)
}
\arguments{
\item{p}{Probability.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\description{
This function computes the quantile of the tail weibull.
}
\examples{
#qtailw()
}
\keyword{weibull}
