% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull-quantile.R
\name{qtailw}
\alias{qtailw}
\title{Quantile function}
\usage{
qtailw(p, threshold, scale, shape)
}
\arguments{
\item{p}{Probability.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\value{
Gives the quantiles of the TailW. The length of the result is determined by the length of x.
}
\description{
This function computes the quantile function of the tailW.
\deqn{Q(p,\alpha, \beta, \nu) = \left(\frac{-\log(1 - p)}{\alpha} + \nu^\beta\right)^{1 / \beta}}
}
\examples{
qtailw(0.5, 1, 1, 1)
}
\references{
Vilardell, Sergi & Serra, Isabel & Abella, Jaume & del Castillo, Joan & Cazorla, Francisco. (2019). Software Timing Analysis for Complex Hardware with Survivability and Risk Analysis. 227-236. <doi:10.1109/ICCD46524.2019.00036>.
}
\keyword{TailW}
