% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG-random.R
\name{rFTG}
\alias{rFTG}
\title{FTG Random Sample Generation}
\usage{
rFTG(n, threshold, scale, shape)
}
\arguments{
\item{n}{Sample size.}

\item{threshold}{Minimum value of the tail.}

\item{scale}{Scale parameter.}

\item{shape}{Shape parameter.}
}
\description{
This function computes n random variates from full-tail gamma with a rejection method.
}
\examples{
x <- rFTG(100, 1, 1, 1)
hist(x, breaks = "FD")
}
\keyword{FTG}
