% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTG-random.R
\name{rFTG}
\alias{rFTG}
\title{Random sample function}
\usage{
rFTG(n, alpha, theta, rho)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{alpha.}

\item{theta}{theta.}

\item{rho}{rho.}
}
\description{
This function computes n random variates from full-tail gamma with a rejection method.
}
\examples{
#rFTG()
}
\keyword{rFTG}
